+++ tests/instruments/thorlabs/test_k10cr2.py
@@ -3,12 +3,12 @@ import struct
 
 from qmi.core.exceptions import QMI_TimeoutException, QMI_InstrumentException
 from qmi.core.transport import QMI_SerialTransport
-from qmi.instruments.thorlabs import Thorlabs_K10Cr1
+from qmi.instruments.thorlabs import Thorlabs_K10Cr2
 from qmi.instruments.thorlabs.apt_packets import _AptMsgHwGetInfo
 
 from tests.patcher import PatcherQmiContext
 
-Thorlabs_K10Cr1.RESPONSE_TIMEOUT = 0.01
+Thorlabs_K10Cr2.RESPONSE_TIMEOUT = 0.01
 # Number of microsteps per degree of rotation.
 MICROSTEPS_PER_DEGREE = 409600.0 / 3.0
 # Internal velocity setting for 1 degree/second.
@@ -17,7 +17,7 @@ VELOCITY_FACTOR = 7329109.0
 ACCELERATION_FACTOR = 1502.0
 
 
-class TestThorlabsK10cr1(unittest.TestCase):
+class TestThorlabsK10cr2(unittest.TestCase):
 
     def setUp(self):
         self._pack = "<l"
@@ -27,7 +27,7 @@ class TestThorlabsK10cr1(unittest.TestCase):
         with unittest.mock.patch(
                 'qmi.instruments.thorlabs.k10crx.create_transport',
                 return_value=self._transport_mock):
-            self.instr: Thorlabs_K10Cr1 = Thorlabs_K10Cr1(PatcherQmiContext(), "instr", "transport_descriptor")
+            self.instr: Thorlabs_K10Cr2 = Thorlabs_K10Cr2(PatcherQmiContext(), "instr", "transport_descriptor")
 
     def tearDown(self):
         self._transport_mock.reset_mock()
@@ -35,10 +35,10 @@ class TestThorlabsK10cr1(unittest.TestCase):
     def test_init(self):
         """Check the __init__ attributes."""
         # Arrange
-        expected_max_velocity = 10
+        expected_max_velocity = 20
         expected_apt_device_address = 0x50
         expected_apt_host_address = 0x01
-        expected_apt_timeout = Thorlabs_K10Cr1.RESPONSE_TIMEOUT
+        expected_apt_timeout = Thorlabs_K10Cr2.RESPONSE_TIMEOUT
         # Assert
         self.assertEqual(expected_max_velocity, self.instr.MAX_VELOCITY)
         self.assertEqual(self.instr._transport, self._transport_mock)
@@ -51,8 +51,8 @@ class TestThorlabsK10cr1(unittest.TestCase):
         # We expect to write MESSAGE_ID 0x0005 (_AptMsgHwReqInfo)
         expected_write = struct.pack(self._pack, 0x0005) + b"P\x01"  # This is 5001 == 0x1389
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo)
-        expected_read = struct.pack(self._pack, 0x0006) + b"\x00" * 2 + b"CR1\0K10" * 12
-        # The request+data has to be 90 bytes long and should include string "K10CR1" at right spot.
+        expected_read = struct.pack(self._pack, 0x0006) + b"\x00" * 2 + b"CR2\0K10" * 12
+        # The request+data has to be 90 bytes long and should include string "K10CR2" at right spot.
         self._transport_mock.read.return_value = expected_read
         self.instr.open()
         # Assert
@@ -67,7 +67,7 @@ class TestThorlabsK10cr1(unittest.TestCase):
         expected_exception = "Expected message type {} not received.".format(_AptMsgHwGetInfo)
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo) but we give 0x0412
         expected_read = struct.pack(self._pack, 0x0412) + b"\x00\x00"
-        self._transport_mock.read.return_value = expected_read + b"K10CR1"
+        self._transport_mock.read.return_value = expected_read + b"K10CR2"
         with self.assertRaises(QMI_TimeoutException) as exc:
             self.instr.open()
 
@@ -98,8 +98,8 @@ class TestThorlabsK10cr1(unittest.TestCase):
         expected_write = struct.pack(self._pack, 0x0005) + b"P\x01"  # This is 5001 == 0x1389
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo)
         expected_read = struct.pack(self._pack, 0x0006) + b"\x81\x5A"
-        # The request+data has to be 90 bytes long and should include string "K10CR1" at right spot.
-        self._transport_mock.read.side_effect = [expected_read, b"CR1\0K10" * 12]
+        # The request+data has to be 90 bytes long and should include string "K10CR2" at right spot.
+        self._transport_mock.read.side_effect = [expected_read, b"CR2\0K10" * 12]
         self.instr.open()
         self.instr.close()
         # Assert
@@ -112,7 +112,7 @@ class TestThorlabsK10cr1(unittest.TestCase):
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo)
         expected_read = struct.pack("<l", 0x0006) + b"\x81\x5A"
         # We expect wrong model message to be
-        exception = "Driver only supports K10CR1 but instrument identifies as 'T10FN2'"
+        exception = "Driver only supports K10CR2 but instrument identifies as 'T10FN2'"
         # The request+data has to be 90 bytes long and should include model type string at right spot.
         self._transport_mock.read.side_effect = [expected_read, b"FN2\0T10" * 12]
         with self.assertRaises(QMI_InstrumentException) as exc:
@@ -134,7 +134,7 @@ class TestThorlabsK10cr1(unittest.TestCase):
         self.assertEqual(expected_exception, str(exc.exception))
 
     def test_open_excepts_with_wrong_data_length(self):
-        """Test opening the instrument excepts when wrong data length is received in _check_k10cr1."""
+        """Test opening the instrument excepts when wrong data length is received in _check_k10cr2."""
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo)
         expected_read = struct.pack("<l", 0x0006) + b"\x81\x5A"
         expected_exception = ("Received partial message (message_id=0x{:04x}, ".format(0x0006) +
@@ -147,7 +147,7 @@ class TestThorlabsK10cr1(unittest.TestCase):
         self.assertEqual(expected_exception, str(exc.exception))
 
 
-class TestThorlabsK10cr1Methods(unittest.TestCase):
+class TestThorlabsK10cr2Methods(unittest.TestCase):
 
     def setUp(self):
         self._pack = "<l"
@@ -155,7 +155,7 @@ class TestThorlabsK10cr1Methods(unittest.TestCase):
         self._displacement = 1 / MICROSTEPS_PER_DEGREE
         self._vel_scaling = VELOCITY_FACTOR
         self._acc_scaling = ACCELERATION_FACTOR
-        self._max_vel = 10
+        self._max_vel = 20
         self._max_acc = 20
         # Mock serial transport
         self._transport_mock = unittest.mock.MagicMock(spec=QMI_SerialTransport)
@@ -164,12 +164,12 @@ class TestThorlabsK10cr1Methods(unittest.TestCase):
         with unittest.mock.patch(
                 'qmi.instruments.thorlabs.k10crx.create_transport',
                 return_value=self._transport_mock):
-            self.instr: Thorlabs_K10Cr1 = Thorlabs_K10Cr1(PatcherQmiContext(), "instr", "transport_descriptor")
+            self.instr: Thorlabs_K10Cr2 = Thorlabs_K10Cr2(PatcherQmiContext(), "instr", "transport_descriptor")
 
         # We expect as response MESSAGE_ID 0x0006 (_AptMsgHwGetInfo) to 'open'
         expected_read = struct.pack("<l", 0x0006)
-        # The request+data has to be 90 bytes long and should include string "K10CR1" at right spot.
-        self._transport_mock.read.return_value = expected_read + self._empty + b"CR1\0K10" * 12
+        # The request+data has to be 90 bytes long and should include string "K10CR2" at right spot.
+        self._transport_mock.read.return_value = expected_read + self._empty + b"CR2\0K10" * 12
         self.instr.open()
         # Clean up the mock for the tests.
         self._transport_mock.reset_mock()
@@ -180,7 +180,7 @@ class TestThorlabsK10cr1Methods(unittest.TestCase):
     def test_get_idn(self):
         """Test the get_idn method."""
         # First two values are hardcoded, and the two latter are based on the standard return value for read in `setUp`
-        expected_idn = ["Thorlabs", "K10CR1", 3232323, "3.2.3"]
+        expected_idn = ["Thorlabs", "K10CR2", 3297859, "3.2.3"]
         # We expect to write MESSAGE_ID 0x0005 (_AptMsgHwReqInfo)
         expected_write = struct.pack(self._pack, 0x0005) + b"P\x01"  # This is 5001 == 0x1389
 
@@ -286,8 +286,8 @@ class TestThorlabsK10cr1Methods(unittest.TestCase):
 
     def test_set_velocity_params_values_out_of_range(self):
         """Test set_velocity_params raises exceptions at invalid values."""
-        invalid_velocities = [0, Thorlabs_K10Cr1.MAX_VELOCITY + 0.1]
-        invalid_accelerations = [0, Thorlabs_K10Cr1.MAX_ACCELERATION + 0.1]
+        invalid_velocities = [0, Thorlabs_K10Cr2.MAX_VELOCITY + 0.1]
+        invalid_accelerations = [0, Thorlabs_K10Cr2.MAX_ACCELERATION + 0.1]
         for max_velocity in invalid_velocities:
             expected1 = f"Invalid value for {max_velocity=}"
             with self.assertRaises(ValueError) as exc1: